+++ tests/test_cli.py
@@ -30,7 +30,7 @@ def test_citation():
 def test_preview():
     with tempfile.TemporaryDirectory() as tmp_dir:
         output = shell_run(
-            f"./bin/sinclair init --output {tmp_dir} && ./bin/sinclair run --output {tmp_dir} -preview"
+            f"./bin/sinclair init --output {tmp_dir} && ./bin/sinclair run --output {tmp_dir} --mode local -preview"
         )
     cmd_line = {
         l.split(":")[0].strip(): l.split(":")[1].strip()
@@ -42,7 +42,7 @@ def test_preview():
 
 def test_forceall():
     output = subprocess.run(
-        "./bin/sinclair run --forceall -preview -profile ci_stub",
+        "./bin/sinclair run --forceall -preview -profile ci_stub --mode local",
         capture_output=True,
         shell=True,
         text=True,
@@ -80,7 +80,7 @@ def test_run_no_init():
     with pytest.raises(Exception) as exc_info:
         with tempfile.TemporaryDirectory() as tmp_dir:
             output = shell_run(
-                f"./bin/sinclair run --output {tmp_dir}",
+                f"./bin/sinclair run --output {tmp_dir} --mode local",
                 check=True,
                 capture_output=True,
             )